Fix hardcoded hour12 and consolidate wallClockToUtc into utils.ts
Remove hour12: true from formatScheduleLocal so locales get their natural
hour cycle (24h for Finland, 12h for US). Move wallClockToUtc from
enrollment.ts to utils.ts as a shared helper — formatScheduleLocal now
calls it instead of duplicating the offset math. Update docs to reflect.
